Adding rundeck module
Along with a lot of other things. Added more orches, added some bootstrap scripts,
added all the needed rundeck pillars and salt files, and also forked the apache
formula so that I could add a param to define the x-forwarded-proto header.

Lastly, did a minor fix with the pki module so that hostnames aren't redundant.